delete source cluster.go
